[ refactor ] Utils.WithDefault, introduced Utils.TypeLits
